Remove test from test case names
