+++ polaris/ocean/suites/nightly.txt
@@ -1,5 +1,5 @@
-ocean/baroclinic_channel/10km/threads_test
-ocean/baroclinic_channel/10km/decomp_test
-ocean/baroclinic_channel/10km/restart_test
+ocean/baroclinic_channel/10km/threads
+ocean/baroclinic_channel/10km/decomp
+ocean/baroclinic_channel/10km/restart
 ocean/inertial_gravity_wave/convergence
 # ocean/manufactured_solution/convergence
