+++ polaris/ocean/suites/pr.txt
@@ -1,6 +1,6 @@
-ocean/baroclinic_channel/10km/threads_test
-ocean/baroclinic_channel/10km/decomp_test
-ocean/baroclinic_channel/10km/restart_test
+ocean/baroclinic_channel/10km/threads
+ocean/baroclinic_channel/10km/decomp
+ocean/baroclinic_channel/10km/restart
 ocean/inertial_gravity_wave/convergence
 ocean/single_column/960km/cvmix
 ocean/single_column/960km/ideal_age
