+++ polaris/ocean/tests/baroclinic_channel/__init__.py
@@ -2,11 +2,11 @@ from polaris import TestGroup
 from polaris.ocean.tests.baroclinic_channel.baroclinic_channel_test_case import (  # noqa: E501
     BaroclinicChannelTestCase,
 )
-from polaris.ocean.tests.baroclinic_channel.decomp_test import DecompTest
+from polaris.ocean.tests.baroclinic_channel.decomp import Decomp
 from polaris.ocean.tests.baroclinic_channel.default import Default
-from polaris.ocean.tests.baroclinic_channel.restart_test import RestartTest
-from polaris.ocean.tests.baroclinic_channel.rpe_test import RpeTest
-from polaris.ocean.tests.baroclinic_channel.threads_test import ThreadsTest
+from polaris.ocean.tests.baroclinic_channel.restart import Restart
+from polaris.ocean.tests.baroclinic_channel.rpe import Rpe
+from polaris.ocean.tests.baroclinic_channel.threads import Threads
 
 
 class BaroclinicChannel(TestGroup):
@@ -26,14 +26,14 @@ class BaroclinicChannel(TestGroup):
                 Default(test_group=self, resolution=resolution))
 
             self.add_test_case(
-                DecompTest(test_group=self, resolution=resolution))
+                Decomp(test_group=self, resolution=resolution))
 
             self.add_test_case(
-                RestartTest(test_group=self, resolution=resolution))
+                Restart(test_group=self, resolution=resolution))
 
             self.add_test_case(
-                ThreadsTest(test_group=self, resolution=resolution))
+                Threads(test_group=self, resolution=resolution))
 
         for resolution in [1., 4., 10.]:
             self.add_test_case(
-                RpeTest(test_group=self, resolution=resolution))
+                Rpe(test_group=self, resolution=resolution))
