+++ polaris/ocean/tests/baroclinic_channel/threads/__init__.py
@@ -3,10 +3,10 @@ from polaris.ocean.tests.baroclinic_channel.forward import Forward
 from polaris.validate import compare_variables
 
 
-class DecompTest(BaroclinicChannelTestCase):
+class Threads(BaroclinicChannelTestCase):
     """
-    A decomposition test case for the baroclinic channel test group, which
-    makes sure the model produces identical results on 1 and 4 cores.
+    A thread test case for the baroclinic channel test group, which makes sure
+    the model produces identical results with 1 and 2 threads.
     """
 
     def __init__(self, test_group, resolution):
@@ -23,25 +23,24 @@ class DecompTest(BaroclinicChannelTestCase):
         """
 
         super().__init__(test_group=test_group, resolution=resolution,
-                         name='decomp_test')
+                         name='threads')
 
-        for procs in [4, 8]:
-            name = f'{procs}proc'
-
+        for openmp_threads in [1, 2]:
+            name = f'{openmp_threads}thread'
             self.add_step(Forward(
-                test_case=self, name=name, subdir=name, ntasks=procs,
-                min_tasks=procs, openmp_threads=1,
+                test_case=self, name=name, subdir=name, ntasks=4,
+                min_tasks=4, openmp_threads=openmp_threads,
                 resolution=resolution, run_time_steps=3))
 
     def validate(self):
         """
         Compare ``temperature``, ``salinity``, ``layerThickness`` and
-        ``normalVelocity`` in the ``4proc`` and ``8proc`` steps with each other
-        and with a baseline if one was provided
+        ``normalVelocity`` in the ``1thread`` and ``2thread`` steps with each
+        other and with a baseline if one was provided
         """
         super().validate()
         variables = ['temperature', 'salinity', 'layerThickness',
                      'normalVelocity']
         compare_variables(test_case=self, variables=variables,
-                          filename1='4proc/output.nc',
-                          filename2='8proc/output.nc')
+                          filename1='1thread/output.nc',
+                          filename2='2thread/output.nc')
